First attempt at solution
Check if the size of the list of numbers is odd or even, then calculate median appropriately. For even, number that is half the size of the list will give index for right-middle number. For odd, the size of list -1 and halved will give index for middle number (-1 to account for 0-indexed)